Handle Scalar IDFs in PciIntervalFromScalarExpr.
Add functionality to return a Constraint Interval for an expression of
the form col IS DISTINCT FROM const.
Also remove an unneccesary check in PexprPredicateCol() for the case of
IS NOT NULL, since it is now supported. But this meant redundant IS NOT
NULL filters, so take care of that also.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>