Enable partition selection when a cast is present
Enabled by including CScalarCmp expressions that contain a CScalarCast
in the types of expressions considered for partition filters. Observe
that if the expression contains a CScalarCast over a CScalarIdent, it
must be preserved all the way to the final plan. That is, we cannot
"peak" and extract the identifer under the cast. For this reason, in
case it is an equality comparison with a cast, levelEqExprs can no
longer be used.

Also, Convert a cast on list part filters to array coerce.

During Expr to DXL translation, construct a CDXLScalarArrayCoerceExpr
operator on top of CDXLScalarPartListValues when a cast is present on
top of the partition key in the partition filter expression.

Also, refactor SplitPartPredicates() to make it easier to read; and
refactor methods around PdxlnListFilterScCmp() by extracting out the
generation of Part Key expression for LIST partition filters.

However, ORCA won't be able to generate a partition filter from an
expression of the form : pk::int8 IS DISTINCT FROM 5. This is because
IDF expressions are handled by the CConstraintInterval framwork which
converts it to a corresponding CScalarCmp + CScalarNullTest. This
framwork cannot preserve cast information on the partkey since it stores
only a CColRef.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>