Remove Broadcast for Inner Nested Loop Join
Prior to this commit, for Inner Join queries with join predicate on the
distribution column produced a plan as follows:

```
-> Gather Motion
   ->  Nested Loop
         Join Filter: distribcolforouter = distribcolforinner
         ->  Broadcast Motion
               ->  Table Scan on outerchild
         ->  Table Scan on innerchild
```

The above plan can be optimized further by removing the broadcast motion
added on top of the outer child. Since the tuples are already
co-located, the broadcast motion is unnecessary and can be dropped.

With this commit, following plan is generated for Inner Join queries
with join predicate on the distribution column
```
-> Gather Motion
   ->  Nested Loop
         Join Filter: distribcolforouter = distribcolforinner
         ->  Table Scan on outerchild
         ->  Table Scan on innerchild
```

Signed-off-by: Abhijit Subramanya <asubramanya@pivotal.io>