Do not push down residual filters in index NLJ
When the join condition of an Index NLJ contains index columns along
with non-index columns (residual filters), ORCA produces a plan with a
filter on the Index Scan. That is, it "pushes" the join filter down to
the Index Scan.

For example:
explain (costs off)
select * from foo, bar where foo.c = bar.c and foo.a = bar.a;

 Gather Motion 3:1  (slice2; segments: 3)
   ->  Nested Loop
         Join Filter: true
         ->  Broadcast Motion 3:3  (slice1; segments: 3)
               ->  Table Scan on bar
         ->  Index Scan using foo_c_bt_idx on foo
               Index Cond: (c = bar.c)
               Filter: (a = bar.a)

Although, this is not incorrect, it means that ORCA misses out on some
optimizations because the join now appears like a cross join. For
example, for the above plan, ORCA could have avoided the Broadcast
motion altogether if both foo and bar were distributed by (a).
Besides, since the pushed filter is not even used in the Index Scan, it
provides negligible performance advantages.

This PR disables this push down of the residual filters.