New interface IPrinter for OsPrint() and DbgPrint() functions.
This commit removes copied implementations of DbgPrint() from a number
of classes and implements it in IPrinter. Note that not all classes that
implement OsPrint/DbgPrint() are converted at this time. This is to keep
the scope of this commit down. Over time, more and more classes should
inherit from IPrinter and use the DbgPrint() defined there.

Also, in some cases (CMemo, CGroup & CGroupExpression), the conversion
was not possible because these classes use CSyncHashTable or CSyncList,
which need to modify the underlying data structure for iteration. This
is incompatible with the const function implementation of OsPrint() - so
leave the non-const implementations in those classes intact for now.