Refactor RewindabilitySpec to improve correlated subquery handling
Previously, ORCA would produces wrong results for some queries with
outer references because of incorrect placement of Spool operator above
the operator with outer references. Sometimes ORCA would error out due
to illegal rescans, because a Spool was missing above a Motion on the
inner side of a correlated join. Also in some cases, ORCA would fall
back due to property mismatch.

This commit fixes many of the above issues by refactoring RewindabilitySpec.

It extends the rewindability type in ORCA's enforcement
framework to include another type: ErtRescannable. It is used to encode
certain nuanced cases that were earlier combined under Rewindable.

The meaning of "Rescannable" is as follows:
require:
  The subtree should be executable multiple times, even though it need
  not return the same results every time. This is required by correlated
  joins from their inner child.
derive:
  The op is not rewindable, but may be rescanned to give a different
  result. (e.g TVF containing a volatile function).

The commit also renames ErtNotRewindable to ErtNone for cases that
are neither rewindable nor rescannable - e.g Motion.

This refactor allows for a more elegant way of handing rewindability
enforcement when outer references are present, instead of repeating an
outer ref check in many operators. Bugs related to this issue resurfaced
many times in the past for different operators, which has been hard to
maintain.

This commit attempts to unify the logic in CRewindabilitySpec &
"materializing" operators like Sort & Spool. The result of these changes
is that now ORCA produces plans for queries for which it previously used
to fall back; it has also removed unnecessary Spools for some cases and
fixed wrong results.