Implement Full Merge Joins in ORCA
This commit implements the infrastructure needed to support creating
plans that use full outer merge joins in ORCA. Part of this
infrastructure was already implemented (e.g CDXLPhysicalMergeJoin),
while this commit fills in the rest with the following changes:

- Add a CPhysicalFullMergeJoin class, including the necessary property
  requirements & derivation methods.
- Add a new xform: CXformImplementFullOuterMergeJoin that forms an
  alternative path for CLogicalFullJoin using CPhysicalFullMergeJoin
- Fix up Expr -> DXL translator with correct alignment of merge join
  conditions with child ops
- Statistics derivation & costing (see notes below for details)

Additional notes about this implementation:

First, this commit supports multi-key merge joins by using a mechanism
(similar to what HashJoin uses) to cache join keys to set the keys in
the mergeclauses. This was possible only after refactors in preceding
commits.

Second, regarding distribution specs, merge join does not require null
collocation for STRICT predicates. This reduces some unnecessary
Motions. Also, merge join also requests singleton distribution
This helps with plans like:
  explain select * from
	 (select sum(a) from foo)   t1(l) full join
	 (select count(c) from bar) t2(r) on (t2.r = t1.l)i;

NB: This commit changes DistributionSpecHashed::FMatch() to better check
compatibility (instead of exact match) for nulls collocation.

Third, this commit disables stats derivation for CLogicalFullOuterJoin.
Instead, use the stats derived from an equivalent UNION ALL expression.

To maintain the ability to enable/disable the ExfExpandFullOuterJoin,
this commit introduces a new traceflag.  Instead of enabling or
disabling the transform directly, it skips the implementation of the
CTEAnchor above the UnionAll, if the trace flag is disabled and the
CTEAnchor op was generated by ExfExpandFullOuterJoin. This way, even if
expansion is disabled by the trace flag, the xform is still triggered,
so that the alternative generated can be used to compute statistics for
the logical merge join group (see CLogicalFullOuterJoin::Esp()).

Finally, IsMergeJoinable was added to types in MDP files. The values
were verified by running on GPDB 6X:
   select oid, *
   from pg_operator
   where
     ((oprcanmerge and not oprcanhash) or (not oprcanmerge and oprcanhash))
       and oprleft = oprright;

But in case of operators like date_eq_timestamp, this doesn't check
pg_operator. Unfortunately, for merge join we can't cast on both sides
because the order spec doesn't maintain expressions, only colrefs.
Luckily date_eq_timestamp is mergejoinable and so are the other such
operators, so no such cast is needed for correctness.

But, just in case, we disallow cases where type(lhs) != type(rhs).

Also, bump ORCA version to 3.59.0.