Skip CXformPushGbBelowSetOp when arity > threshold
In case of queries containing a UNION over a large number of children,
CXformPushGbBelowSetOp pushes the Group By op to each child, creating
that many new groups in the memo. While not a problem by itself, it
exponentially increases the number of bindings for other xforms that use
CPatternTree to extract it. When extracting the bindings, each child of
UNION can now be selected with and without the group by. The higher the
number of such children, the more the number of possibilities, the
higher the optimization time.

An example of xform that is affected: CXformLeftSemiApply2LeftSemiJoin.

To prevent this exponential increase in binding extraction, restrict
when the CXformPushGbBelowSetOp may be executed by a configurable
parameter.