Remove unused 'thread-safe' flags from memory pools.
They were not used for anything. It's a nice idea to differentiate
between memory pools that need to be thread-safe and those that do not,
but given that the flag has not been used for anything for ages, it seems
likely that it has bit-rotted and we're not actually marking things
correctly. So seems better to keep things simple and remove it.