Remove CMemoryPoolStack
We only used the CMemoryPoolStack when creating the scheduler context.
The CMemoryPoolStack was a memory pool that only freed memory when the
pool was destroyed, not when we explicitly called free(). This
functionality is no longer used with the introduction of
CMemoryPoolPalloc, and for consistency we now just use
CMemoryPoolTracker if the Palloc pool is not being used.