Allow only equality comparisons for Dynamic Partition Elimination
This commit only allows equality comparisons when doing dynamic
partition elimination. It will be the default behavior moving forward.

Non-equality predicates for dynamic partition elimination is currently
expensive to execute since the executor must iterate over all the
partition rules for each row from its subtree and execute the
non-equality predicate. So for cases where there are a large number of
rows and/or partitions, this process of selecting the partition may
outweigh the savings gained by skipping the eliminated partitions.

The commit fixes the erroneous logic for removing "IS NOT NULL" exprs

They should only be removed if the selected partition expressions are
strict. Also add assert checks for certain assumptions made to test this
logic.

The commit also includes some minor refactors and removal of dead code.

MDP changes:

* Only plan size or cost changes (due to removal of IS NOT NULL
  predicates)

    data/dxl/minidump/DPE-SemiJoin.mdp
    data/dxl/minidump/IndexApply-PartKey-Is-IndexKey.mdp
    data/dxl/minidump/IndexApply-PartResolverExpand.mdp
    data/dxl/minidump/PartTbl-CSQ-PartKey.mdp
    data/dxl/minidump/SpoolShouldInvalidateUnresolvedDynamicScans.mdp
    data/dxl/minidump/IndexApply-Heterogeneous-BothSidesPartitioned.mdp

* Trace flag is added to preserve old behavior since they test specific
  scenarios which are now disabled

    data/dxl/minidump/DPE-with-unsupported-pred.mdp
    data/dxl/minidump/NLJ-Broadcast-DPE-Outer-Child.mdp
    data/dxl/minidump/PartTbl-IDFNull.mdp
    data/dxl/minidump/PartTbl-RangeJoinPred.mdp

* Addition of IS NOT NULL predicate due to stricter checks

    data/dxl/minidump/PartTbl-CSQ-PartKey.mdp

Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>
Co-authored-by: Ashuka Xue <axue@pivotal.io>